Split time class in separate header file
