+++ include/solid_mechanics/mf_elasticity.h
@@ -64,6 +64,7 @@ static const unsigned int debug_level = 0;
 #include <adapter/precice_adapter.h>
 #include <base/fe_integrator.h>
 #include <base/q_equidistant.h>
+#include <base/time_handler.h>
 #include <base/version.h>
 #include <cases/case_base.h>
 #include <parameter/parameter_handling.h>
@@ -82,57 +83,6 @@ using namespace dealii;
 namespace FSI
 {
   using namespace dealii;
-
-  // @sect3{Time class}
-
-  // A simple class to store time data. Its functioning is transparent so no
-  // discussion is necessary. For simplicity we assume a constant time step
-  // size.
-  class Time
-  {
-  public:
-    Time(const double time_end, const double delta_t)
-      : timestep(0)
-      , time_current(0.0)
-      , time_end(time_end)
-      , delta_t(delta_t)
-    {}
-
-    virtual ~Time() = default;
-
-    double
-    current() const
-    {
-      return time_current;
-    }
-    double
-    end() const
-    {
-      return time_end;
-    }
-    double
-    get_delta_t() const
-    {
-      return delta_t;
-    }
-    unsigned int
-    get_timestep() const
-    {
-      return timestep;
-    }
-    void
-    increment()
-    {
-      time_current += delta_t;
-      ++timestep;
-    }
-
-  private:
-    unsigned int timestep;
-    double       time_current;
-    const double time_end;
-    const double delta_t;
-  };
 
   // The Solid class is the central class.
   template <int dim, typename Number>
